Move ChangesetCounts and their calculation to enterprise/a8n
